Rename Path to TasksPath and Remove side effect
