--- app/src/main/java/com/codesoom/assignment/models/Path.java
@@ -1,24 +1,0 @@
-package com.codesoom.assignment.models;
-
-import java.util.OptionalLong;
-
-public class Path {
-    private String path;
-    private String target = "/tasks/";
-    public Path(String path) {
-        this.path = path;
-    }
-
-    public Long extractNumber() {
-        path = path.replace(target, "");
-        return OptionalLong.of(Long.parseLong(path)).getAsLong();
-    }
-    public boolean checkUrl(){
-        return path.equals("/") || path.startsWith("/tasks");
-    }
-
-    public boolean hasNumberParameter() {
-        path = path.replace(target, "");
-        return path.matches("^[0-9]+$");
-    }
-}
