+++ app/src/main/java/com/codesoom/assignment/models/TasksPath.java
@@ -1,0 +1,24 @@
+package com.codesoom.assignment.models;
+
+import java.util.OptionalLong;
+
+public class TasksPath {
+    private final String path;
+    private static final String TASKS = "/tasks/";
+    public TasksPath(String path) {
+        this.path = path;
+    }
+
+    public Long extractNumber() {
+        String replace = path.replace(TASKS, "");
+        return OptionalLong.of(Long.parseLong(replace)).getAsLong();
+    }
+    public boolean checkUrl(){
+        return path.equals("/") || path.startsWith("/tasks");
+    }
+
+    public boolean hasNumberParameter() {
+        String replace = path.replace(TASKS, "");
+        return replace.matches("^[0-9]+$");
+    }
+}
